Extract Assign to separate struct
